Check module enabled for cities migration

+++ src/Database/Migrations/2020_07_07_055725_create_cities_table.php
@@ -13,18 +13,20 @@ class CreateCitiesTable extends Migration
 	 */
 	public function up()
 	{
-		Schema::create(config('world.migrations.cities.table_name'), function (Blueprint $table) {
-			$table->id();
-			$table->foreignId('country_id');
-			$table->foreignId('state_id');
-			$table->string('name');
+		if(config('world.modules.cities')){
+			Schema::create(config('world.migrations.cities.table_name'), function (Blueprint $table) {
+				$table->id();
+				$table->foreignId('country_id');
+				$table->foreignId('state_id');
+				$table->string('name');
 
-			foreach (config('world.migrations.cities.optional_fields') as $field => $value) {
-				if ($value['required']) {
-					$table->string($field, $value['length'] ?? null);
+				foreach (config('world.migrations.cities.optional_fields') as $field => $value) {
+					if ($value['required']) {
+						$table->string($field, $value['length'] ?? null);
+					}
 				}
-			}
-		});
+			});
+		}
 	}
 
 	/**
@@ -34,6 +36,8 @@ class CreateCitiesTable extends Migration
 	 */
 	public function down()
 	{
-		Schema::dropIfExists(config('world.migrations.cities.table_name'));
+		if(config('world.modules.cities')){
+			Schema::dropIfExists(config('world.migrations.cities.table_name'));
+		}
 	}
 }
